Code review feedback pt2
