+++ src/Compilers/Core/Portable/PublicAPI.Unshipped.txt
@@ -1,3 +1,2 @@
 Microsoft.CodeAnalysis.CommandLineSourceFile.CommandLineSourceFile(string path, bool isScript, bool isInputRedirected) -> void
 Microsoft.CodeAnalysis.CommandLineSourceFile.IsInputRedirected.get -> bool
-const Microsoft.CodeAnalysis.LanguageNames.TypeScript = "TypeScript" -> string
