+++ src/EditorFeatures/Core/ExternalAccess/VSTypeScript/Api/VSTypeScriptGlyph.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript
+namespace Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript.Api
 {
     internal enum VSTypeScriptGlyph
     {
