+++ src/EditorFeatures/Core/ExternalAccess/VSTypeScript/VSTypeScriptGlyphHelpers.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript.Api;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.ExternalAccess.VSTypeScript
